# static/swagger.json
{"openapi": "3.0.0", "info": {"title": "Flask Authentication API", "description": "API for user authentication with email verification", "version": "1.0.0"}, "servers": [{"url": "http://localhost:5000", "description": "Development server"}], "components": {"securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}}}, "tags": [{"name": "Authentication", "description": "Authentication operations"}, {"name": "User", "description": "User operations"}, {"name": "Email Verification", "description": "Email verification operations"}], "paths": {"/register": {"post": {"tags": ["Authentication"], "summary": "Register a new user", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string", "example": "testuser"}, "email": {"type": "string", "format": "email", "example": "user@example.com"}, "password": {"type": "string", "format": "password", "example": "password123"}}, "required": ["username", "email", "password"]}}}}, "responses": {"201": {"description": "User created successfully"}, "400": {"description": "Bad request - missing fields or user already exists"}}}}, "/login": {"post": {"tags": ["Authentication"], "summary": "Login to get an access token", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "properties": {"username": {"type": "string", "example": "testuser"}, "password": {"type": "string", "format": "password", "example": "password123"}}, "required": ["username", "password"]}}}}, "responses": {"200": {"description": "Login successful", "content": {"application/json": {"schema": {"type": "object", "properties": {"message": {"type": "string", "example": "Login successful"}, "token": {"type": "string", "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."}, "expires_in": {"type": "integer", "example": 86400}, "email_verified": {"type": "boolean", "example": true}}}}}}, "401": {"description": "Invalid credentials"}}}}, "/verify-email/{token}": {"get": {"tags": ["Email Verification"], "summary": "Verify email address", "parameters": [{"name": "token", "in": "path", "required": true, "schema": {"type": "string"}, "description": "Email verification token"}], "responses": {"200": {"description": "Email verified successfully"}, "400": {"description": "Invalid or expired token"}, "404": {"description": "User not found"}}}}, "/resend-verification": {"post": {"tags": ["Email Verification"], "summary": "Resend verification email", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Verification email sent"}, "400": {"description": "Email already verified"}, "401": {"description": "Unauthorized"}}}}, "/user": {"get": {"tags": ["User"], "summary": "Get current user profile", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "User profile", "content": {"application/json": {"schema": {"type": "object", "properties": {"id": {"type": "integer", "example": 1}, "username": {"type": "string", "example": "testuser"}, "email": {"type": "string", "format": "email", "example": "user@example.com"}, "email_verified": {"type": "boolean", "example": true}, "created_at": {"type": "string", "format": "date-time", "example": "2023-05-20T12:34:56.789012"}}}}}}, "401": {"description": "Unauthorized"}}}}, "/protected-verified": {"get": {"tags": ["User"], "summary": "Access protected resource (requires verified email)", "security": [{"bearerAuth": []}], "responses": {"200": {"description": "Access granted"}, "401": {"description": "Unauthorized"}, "403": {"description": "Email verification required"}}}}, "/test-verification/{username}": {"get": {"tags": ["Testing"], "summary": "Generate test verification email for a user", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string"}, "description": "Username of the user"}], "responses": {"200": {"description": "Test verification email sent"}, "404": {"description": "User not found"}}}}}}